增加加工单的一些单测 改了一些bug

+++ alerp-spring-boot/src/main/java/edu/nju/alerp/dto/ProcessingOrderProductDTO.java
@@ -12,5 +12,5 @@ public class ProcessingOrderProductDTO {
     private Integer productId;
     private String specification;
     private Integer quantity;
-    private Integer expectedWeight;
+    private Double expectedWeight;
 }
